Fix based on Shinji's review
- Separate collecting statistics part and parallelization part
